Move AggregateQuerysetMixin into generic module
